refactor: move emission into entity
